# artifacts/pos-system/src/backend/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Restaurant ERP + POS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tables:
    get:
      operationId: getTables
      tags: [tables]
      summary: List all tables
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
            enum: [available, occupied, reserved, null]
      responses:
        "200":
          description: Tables list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Table"
    post:
      operationId: createTable
      tags: [tables]
      summary: Create a new table
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TableInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Table"
  /tables/{id}:
    get:
      operationId: getTable
      tags: [tables]
      summary: Get a table by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Table
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Table"
    patch:
      operationId: updateTable
      tags: [tables]
      summary: Update a table status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TableUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Table"
    delete:
      operationId: deleteTable
      tags: [tables]
      summary: Delete a table
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ===== ORDERS =====
...
components:
  schemas:
...
    Table:
      type: object
      properties:
        id:
          type: integer
        number:
          type: integer
        name:
          type: string
        capacity:
          type: integer
        status:
          type: string
          enum: [available, occupied, reserved]
        section:
          type: ["string", "null"]
        currentOrderId:
          type: ["integer", "null"]
        createdAt:
          type: string
      required:
        - id
        - number
        - name
        - capacity
        - status
        - createdAt
    TableInput:
      type: object
      properties:
        number:
          type: integer
        name:
          type: string
        capacity:
          type: integer
        section:
          type: ["string", "null"]
      required:
        - number
        - name
        - capacity
    TableUpdate:
      type: object
      properties:
        number:
          type: integer
        name:
          type: string
        capacity:
          type: integer
        status:
          type: string
          enum: [available, occupied, reserved]
        section:
          type: ["string", "null"]

    # ===== ORDER SCHEMAS =====
...